Added a health drop for perfect rounds, the parachute is a work of art. Adjusted spawn rate. Slowed down fire rate.

+++ Assets/Scripts/EnemyCollided.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class EnemyCollided : MonoBehaviour {
 
-    public float collideAttack;
+	public int collideAttack;
     public AudioSource audio;
 	// Use this for initialization
 	void Start () {
